Split CurrentPlayer string functions to cpprint.py
