+++ eudplib/eudlib/stringf/eudprint.py
@@ -29,181 +29,8 @@ from eudplib import (
     utils as ut,
 )
 
-from ..memiof import CPByteWriter, f_setcurpl, f_dwread_epd
 from .rwcommon import br1, bw1
-from .cpstr import _s2b, CPString
 from .dbstr import DBString
-from ..eudarray import EUDArray
-from ..utilf import f_getuserplayerid
-
-cw = CPByteWriter()
-player_colors = "\x08\x0E\x0F\x10\x11\x15\x16\x17\x18\x19\x1B\x1C\x1D\x1E\x1F"
-Color = EUDArray([ut.EPD(c.Db(ut.u2b(p) + b"\0")) for p in player_colors])
-prevcp = c.EUDVariable()
-
-
-def PColor(i):
-    return epd2s(Color[i])
-
-
-def PName(x):
-    if ut.isUnproxyInstance(x, type(c.P1)):
-        x = c.EncodePlayer(x)
-        if x == c.EncodePlayer(c.CurrentPlayer):
-            x = prevcp
-    return ptr2s(0x57EEEB + 36 * x)
-
-
-@c.EUDFunc
-def _addstr_cp():
-    b = c.EUDVariable()
-    if cs.EUDInfLoop()():
-        c.SetVariables(b, br1.readbyte())
-        cs.EUDBreakIf(b == 0)
-        cw.writebyte(b)
-    cs.EUDEndInfLoop()
-
-    cw.flushdword()
-
-
-@c.EUDFunc
-def f_addstr_cp(src):
-    """Print string as string to CurrentPlayer
-
-    :param src: Source address (Not EPD player)
-    """
-    br1.seekoffset(src)
-    _addstr_cp()
-
-
-@c.EUDFunc
-def f_addstr_cp_epd(epd):
-    """Print string as string to CurrentPlayer
-
-    :param epd: EPD player of Source address
-    """
-    br1.seekepd(epd)
-    _addstr_cp()
-
-
-@c.EUDFunc
-def f_adddw_cp(number):
-    """Print number as string to CurrentPlayer.
-
-    :param number: DWORD to print
-    """
-    skipper = [c.Forward() for _ in range(9)]
-    ch = [0] * 10
-
-    # Get digits
-    for i in range(10):
-        number, ch[i] = c.f_div(number, 10)
-        if i != 9:
-            cs.EUDJumpIf(number == 0, skipper[i])
-
-    # print digits
-    for i in range(9, -1, -1):
-        if i != 9:
-            skipper[i] << c.NextTrigger()
-        cw.writebyte(ch[i] + b"0"[0])
-
-    cw.flushdword()
-
-
-@c.EUDFunc
-def f_addptr_cp(number):
-    """Print number as string to CurrentPlayer.
-
-    :param number: DWORD to print
-    """
-    digit = [c.EUDLightVariable() for _ in range(8)]
-    cs.DoActions([
-        [digit[i].SetNumber(0) for i in range(8)],
-        c.SetDeaths(c.CurrentPlayer, c.SetTo, ut.b2i4(b"0000"), 0),
-    ])
-
-    def f(x):
-        t = x % 16
-        q, r = divmod(t, 4)
-        return 2 ** (r + 8 * (3 - q))
-
-    for i in range(31, -1, -1):
-        c.RawTrigger(
-            conditions=number.AtLeast(2 ** i),
-            actions=[
-                number.SubtractNumber(2 ** i),
-                digit[i // 4].AddNumber(2 ** (i % 4)),
-                c.SetDeaths(c.CurrentPlayer, c.Add, f(i), 0),
-            ],
-        )
-        if i % 16 == 0:
-            for j in range(4):
-                c.RawTrigger(
-                    conditions=digit[j + 4 * (i // 16)].AtLeast(10),
-                    actions=c.SetDeaths(
-                        c.CurrentPlayer, c.Add,
-                        (b"A"[0] - b":"[0]) * (256 ** (3 - j)), 0
-                    ),
-                )
-            cs.DoActions([
-                c.AddCurrentPlayer(1),
-                [
-                    c.SetDeaths(c.CurrentPlayer, c.SetTo, ut.b2i4(b"0000"), 0)
-                    if i == 16
-                    else []
-                ],
-            ])
-
-
-_constcpstr_dict = dict()
-
-
-class ptr2s:
-    def __init__(self, value):
-        self._value = value
-
-
-class epd2s:
-    def __init__(self, value):
-        self._value = value
-
-
-def f_cpstr_print(*args):
-    """Print multiple string / number to CurrentPlayer.
-
-    :param args: Things to print
-
-    """
-    args = ut.FlattenList(args)
-    for arg in args:
-        if ut.isUnproxyInstance(arg, str):
-            arg = ut.u2utf8(arg)
-        elif ut.isUnproxyInstance(arg, int):
-            arg = ut.u2b(str(arg & 0xFFFFFFFF))
-        if ut.isUnproxyInstance(arg, bytes):
-            key = _s2b(arg)
-            if key not in _constcpstr_dict:
-                _constcpstr_dict[key] = CPString(arg)
-            arg = _constcpstr_dict[key]
-        if ut.isUnproxyInstance(arg, CPString):
-            arg.Display()
-        elif ut.isUnproxyInstance(arg, ptr2s):
-            f_addstr_cp(arg._value)
-        elif ut.isUnproxyInstance(arg, epd2s):
-            f_addstr_cp_epd(arg._value)
-        elif ut.isUnproxyInstance(arg, DBString):
-            f_addstr_cp_epd(ut.EPD(arg.GetStringMemoryAddr()))
-        elif ut.isUnproxyInstance(arg, c.EUDVariable) or c.IsConstExpr(arg):
-            f_adddw_cp(arg)
-        elif ut.isUnproxyInstance(arg, hptr):
-            f_addptr_cp(arg._value)
-        else:
-            raise ut.EPError(
-                "Object with unknown parameter type %s given to f_cpprint."
-                % type(arg)
-            )
-    # EOS
-    # cs.DoActions(c.SetDeaths(c.CurrentPlayer, c.SetTo, 0, 0))
 
 
 @c.EUDFunc
@@ -292,6 +119,16 @@ def f_dbstr_addptr(dst, number):
     return dst
 
 
+class ptr2s:
+    def __init__(self, value):
+        self._value = value
+
+
+class epd2s:
+    def __init__(self, value):
+        self._value = value
+
+
 class hptr:
     def __init__(self, value):
         self._value = value
@@ -352,26 +189,3 @@ def f_simpleprint(*args, spaced=True):
     # Print
     f_dbstr_print(_printf_buffer, *args)
     _printf_buffer.Display()
-
-
-@c.EUDTypedFunc([c.TrgPlayer])
-def f_raise_CCMU(player):
-    orignextptr = f_dwread_epd(ut.EPD(0x628438))
-    cs.DoActions([
-        c.SetMemory(0x628438, c.SetTo, 0),
-        c.CreateUnit(1, 0, 64, player),
-        c.SetMemory(0x628438, c.SetTo, orignextptr),
-    ])
-
-
-def f_eprintln(*args):
-    f_raise_CCMU(c.CurrentPlayer)
-    localcp = f_getuserplayerid()
-    if cs.EUDIf()(c.Memory(0x6509B0, c.Exactly, localcp)):
-        f_setcurpl(ut.EPD(0x640B60 + 218 * 12))
-        f_cpstr_print(*args)
-        cs.DoActions([
-            c.SetDeaths(c.CurrentPlayer, c.SetTo, 0, 0),
-            c.SetCurrentPlayer(localcp),
-        ])
-    cs.EUDEndIf()
